[SharedModel] Build clean VS, cmake/clang, cmake/VS, and cmake/emscripten
