+++ source/shared/cpp/ObjectModel/DateTimePreparser.cpp
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 #include "pch.h"
 
-#if defined(__ANDROID__) || (__APPLE__) || (__linux__)
+#if defined(__ANDROID__) || (__APPLE__) || (__linux__) || (__EMSCRIPTEN__)
 #define LOCALTIME(X, Y) (nullptr == localtime_r(Y, X))
 #else
 #define LOCALTIME(X, Y) localtime_s(X, Y)
