+++ source/shared/cpp/ObjectModel/HostConfig.h
@@ -21,12 +21,12 @@ namespace AdaptiveSharedNamespace
         void SetFontSize(TextSize size, unsigned int value);
 
     private:
-        // UINT_MAX used to check if value was defined
-        unsigned int _small = UINT_MAX;
-        unsigned int _default = UINT_MAX;
-        unsigned int _medium = UINT_MAX;
-        unsigned int _large = UINT_MAX;
-        unsigned int _extraLarge = UINT_MAX;
+        // std::numeric_limits<unsigned int>::max() used to check if value was defined
+        unsigned int _small = std::numeric_limits<unsigned int>::max();
+        unsigned int _default = std::numeric_limits<unsigned int>::max();
+        unsigned int _medium = std::numeric_limits<unsigned int>::max();
+        unsigned int _large = std::numeric_limits<unsigned int>::max();
+        unsigned int _extraLarge = std::numeric_limits<unsigned int>::max();
     };
 
     class FontWeightsConfig
@@ -43,10 +43,10 @@ namespace AdaptiveSharedNamespace
         void SetFontWeight(TextWeight weight, unsigned int value);
 
     private:
-        // UINT_MAX used to check if value was defined
-        unsigned int _lighter = UINT_MAX;
-        unsigned int _default = UINT_MAX;
-        unsigned int _bolder = UINT_MAX;
+        // std::numeric_limits<unsigned int>::max() used to check if value was defined
+        unsigned int _lighter = std::numeric_limits<unsigned int>::max();
+        unsigned int _default = std::numeric_limits<unsigned int>::max();
+        unsigned int _bolder = std::numeric_limits<unsigned int>::max();
     };
 
     struct FontTypeDefinition
